Updated doc and added tests for Hazelcast embedded
